Added separate file for database functions
